# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chroma - Professional Filmmaker Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Browse/search filmmakers
      parameters:
        - name: profession
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: skills
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [featured, newest, most_viewed]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Get a user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # ── Videos ─────────────────────────────────────────────────────────────────
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List/search videos
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, most_viewed, featured]
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoList"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Create a new video entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoInput"
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a single video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
    patch:
      operationId: updateVideo
      tags: [videos]
      summary: Update a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoUpdate"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
    delete:
      operationId: deleteVideo
      tags: [videos]
      summary: Delete a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /videos/{id}/like:
    post:
      operationId: likeVideo
      tags: [videos]
      summary: Like a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /videos/{id}/comments:
    get:
      operationId: listComments
      tags: [videos]
      summary: Get comments for a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentList"
    post:
      operationId: createComment
      tags: [videos]
      summary: Post a comment on a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ── Follows ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, name, profession, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        name:
          type: string
        profession:
          type: string
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        socialLinks:
          type: ["object", "null"]
        isFollowing:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        videoCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UserProfile:
      type: object
      required: [user, videos]
      properties:
        user:
          $ref: "#/components/schemas/User"
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        credits:
          type: array
          items:
            type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        profession:
          type: string
        location:
          type: string
        bio:
          type: string
        website:
          type: string
        avatarUrl:
          type: string
        coverUrl:
          type: string
        skills:
          type: array
          items:
            type: string
        socialLinks:
          type: object
    UserList:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer

    # ── Video schemas ─────────────────────────────────────────────────────────
    Video:
      type: object
      required: [id, userId, title, privacy, viewCount, likeCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        privacy:
          type: string
          enum: [public, private, password_protected]
        tags:
          type: array
          items:
            type: string
        credits:
          type: ["string", "null"]
        viewCount:
          type: integer
        likeCount:
          type: integer
        isLiked:
          type: boolean
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    VideoInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        privacy:
          type: string
          enum: [public, private, password_protected]
        tags:
          type: array
          items:
            type: string
        credits:
          type: string
    VideoUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
        privacy:
          type: string
          enum: [public, private, password_protected]
        tags:
          type: array
          items:
            type: string
        credits:
          type: string
    VideoList:
      type: object
      required: [videos, total]
      properties:
        videos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        total:
          type: integer
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
    Comment:
      type: object
      required: [id, videoId, userId, body, createdAt]
      properties:
        id:
          type: integer
        videoId:
          type: integer
        userId:
          type: integer
        body:
          type: string
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    CommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
    CommentList:
      type: object
      required: [comments]
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"

    # ── Follow schemas ────────────────────────────────────────────────────────
...