# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chroma - Professional Filmmaker Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateSocialLinks } from './userUpdateSocialLinks';

export interface UserUpdate {
  username?: string;
  name?: string;
  profession?: string;
  location?: string;
  bio?: string;
  website?: string;
  avatarUrl?: string;
  coverUrl?: string;
  skills?: string[];
  socialLinks?: UserUpdateSocialLinks;
}