# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chroma - Professional Filmmaker Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';
import type { VideoPrivacy } from './videoPrivacy';

export interface Video {
  id: number;
  userId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  privacy: VideoPrivacy;
  tags?: string[];
  /** @nullable */
  credits?: string | null;
  viewCount: number;
  likeCount: number;
  isLiked?: boolean;
  user?: User;
  createdAt: Date;
}